NXP-25543: move collection operation to nuxeo-automation-feature + clean up dependencies
tot
