+++ nuxeo-dm/nuxeo-platform-webapp-types/src/main/resources/web/nuxeo.war/icons/collection.png
