+++ nuxeo-dm/nuxeo-platform-webapp-types/src/main/resources/web/nuxeo.war/icons/pin_100.png
