+++ nuxeo-features/nuxeo-automation/nuxeo-automation-features/pom.xml
@@ -89,6 +89,11 @@
       <type>test-jar</type>
       <scope>test</scope>
     </dependency>
+    <!-- for collections operations -->
+    <dependency>
+      <groupId>org.nuxeo.ecm.platform</groupId>
+      <artifactId>nuxeo-platform-collections-core</artifactId>
+    </dependency>
     <!-- for principal related operations -->
     <dependency>
       <groupId>org.nuxeo.ecm.platform</groupId>
@@ -210,6 +215,20 @@
       <groupId>org.nuxeo.ecm.platform</groupId>
       <artifactId>nuxeo-platform-test</artifactId>
       <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.nuxeo.ecm.platform</groupId>
+      <artifactId>nuxeo-platform-userworkspace-types</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.nuxeo.ecm.platform</groupId>
+      <artifactId>nuxeo-platform-userworkspace-core</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.nuxeo.ecm.core</groupId>
+      <artifactId>nuxeo-core-query</artifactId>
     </dependency>
     <dependency>
       <groupId>org.nuxeo.runtime</groupId>
