+++ nuxeo-features/nuxeo-platform-collections/nuxeo-platform-collections-core/src/main/resources/META-INF/MANIFEST.MF
@@ -6,22 +6,18 @@ Bundle-Version: 0.0.0.SNAPSHOT
 Bundle-Name: Nuxeo ECM Collections
 Bundle-ManifestVersion: 2
 Nuxeo-Component: OSGI-INF/collection-core-types-contrib.xml,
- OSGI-INF/collection-ui-types-contrib.xml,
  OSGI-INF/collection-lifecycle-contrib.xml,
  OSGI-INF/collection-pageprovider-contrib.xml,
  OSGI-INF/collection-service.xml,
  OSGI-INF/collection-adapter-contrib.xml,
  OSGI-INF/collection-security-contrib.xml,
- OSGI-INF/collection-operations-contrib.xml,
  OSGI-INF/collection-event-handlers-contrib.xml,
  OSGI-INF/collection-workmanager-contrib.xml,
  OSGI-INF/collection-audit-contrib.xml,
  OSGI-INF/collection-versioning-contrib.xml,
  OSGI-INF/favorites-core-types-contrib.xml,
- OSGI-INF/favorites-ui-types-contrib.xml,
  OSGI-INF/favorites-service.xml,
  OSGI-INF/favorites-lifecycle-contrib.xml,
  OSGI-INF/json-enrichers-contrib.xml
 Bundle-SymbolicName: org.nuxeo.ecm.platform.collections.core;singleton:=true
 Bundle-RequiredExecutionEnvironment: JavaSE-1.6
-Require-Bundle: org.nuxeo.ecm.platform.collections.api
