+++ nuxeo-features/nuxeo-platform-collections/nuxeo-platform-collections-core/src/main/resources/OSGI-INF/collection-core-types-contrib.xml
@@ -3,7 +3,6 @@
 <component name="org.nuxeo.ecm.collections.schemas">
 
   <require>org.nuxeo.ecm.core.CoreExtensions</require>
-  <require>org.nuxeo.ecm.tags.schemas</require>
 
   <extension target="org.nuxeo.ecm.core.schema.TypeService" point="schema">
 
@@ -35,7 +34,6 @@
       <facet name="Versionable"/>
       <facet name="Collection" />
       <facet name="NotCollectionMember" />
-      <facet name="NXTag" />
       <schema name="dublincore" />
       <schema name="common" />
     </doctype>
