--- nuxeo-features/nuxeo-platform-collections/nuxeo-platform-collections-core/src/main/resources/OSGI-INF/collection-operations-contrib.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0"?>
-<component name="org.nuxeo.collections.operations" version="1.0">
-  <extension target="org.nuxeo.ecm.core.operation.OperationServiceComponent"
-             point="operations">
-
-    <operation class="org.nuxeo.ecm.collections.core.automation.SuggestCollectionEntry" />
-    <operation class="org.nuxeo.ecm.collections.core.automation.CreateCollectionOperation" />
-    <operation class="org.nuxeo.ecm.collections.core.automation.AddToCollectionOperation" />
-    <operation class="org.nuxeo.ecm.collections.core.automation.RemoveFromCollectionOperation" />
-    <operation class="org.nuxeo.ecm.collections.core.automation.RemoveFromFavoritesOperation" />
-    <operation class="org.nuxeo.ecm.collections.core.automation.AddToFavoritesOperation" />
-    <operation class="org.nuxeo.ecm.collections.core.automation.GetCollectionsOperation" />
-    <operation class="org.nuxeo.ecm.collections.core.automation.GetDocumentsFromCollectionOperation" />
-    <operation class="org.nuxeo.ecm.collections.core.automation.GetDocumentsFromFavoritesOperation" />
-    <operation class="org.nuxeo.ecm.collections.core.automation.FetchFavorites" />
-    <operation class="org.nuxeo.ecm.collections.core.automation.MoveCollectionMemberOperation" />
-
-  </extension>
-</component>
