--- nuxeo-features/nuxeo-platform-collections/nuxeo-platform-collections-core/src/main/resources/OSGI-INF/collection-ui-types-contrib.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0"?>
-
-<component name="org.nuxeo.ecm.collections.uiTypes">
-
-  <extension target="org.nuxeo.ecm.platform.types.TypeService" point="types">
-
-    <type id="Collections">
-      <label>Collections</label>
-      <description></description>
-      <default-view>view_documents</default-view>
-      <icon>/icons/collection.png</icon>
-      <bigIcon>/icons/collection_100.png</bigIcon>
-      <category>Collaborative</category>
-      <description>Folder.description</description>
-      <default-view>view_documents</default-view>
-      <layouts mode="any">
-        <layout>heading</layout>
-      </layouts>
-      <layouts mode="edit">
-        <layout>heading</layout>
-        <layout>dublincore</layout>
-      </layouts>
-      <contentViews category="content">
-        <contentView>document_content</contentView>
-      </contentViews>
-      <contentViews category="trash_content">
-        <contentView showInExportView="false">
-          document_trash_content
-        </contentView>
-      </contentViews>
-    </type>
-
-    <type id="Collection">
-      <label>Collection</label>
-      <description></description>
-      <default-view>view_documents</default-view>
-      <icon>/icons/collection.png</icon>
-      <bigIcon>/icons/collection_100.png</bigIcon>
-      <category>Collaborative</category>
-      <description>Folder.description</description>
-      <default-view>view_documents</default-view>
-      <layouts mode="any">
-        <layout>heading</layout>
-      </layouts>
-      <layouts mode="edit">
-        <layout>heading</layout>
-        <layout>dublincore</layout>
-      </layouts>
-      <contentViews category="collectionContent">
-        <contentView showInExportView="false">collection_content_contentview</contentView>
-      </contentViews>
-    </type>
-
-  </extension>
-
-</component>
