--- nuxeo-features/nuxeo-platform-collections/nuxeo-platform-collections-core/src/main/resources/OSGI-INF/favorites-ui-types-contrib.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0"?>
-
-<component name="org.nuxeo.ecm.favorites.uiTypes">
-
-  <require>org.nuxeo.ecm.collections.uiTypes</require>
-
-  <extension target="org.nuxeo.ecm.platform.types.TypeService" point="types">
-
-    <type id="Favorites">
-      <label>Favorites</label>
-      <description></description>
-      <default-view>view_documents</default-view>
-      <icon>/icons/pin.png</icon>
-      <bigIcon>/icons/pin_100.png</bigIcon>
-      <description>Folder.description</description>
-      <default-view>view_documents</default-view>
-      <layouts mode="any">
-        <layout>heading</layout>
-      </layouts>
-      <layouts mode="edit">
-        <layout>heading</layout>
-        <layout>dublincore</layout>
-      </layouts>
-      <contentViews category="collectionContent">
-        <contentView showInExportView="false">collection_content_contentview</contentView>
-      </contentViews>
-    </type>
-
-  </extension>
-
-</component>
