+++ nuxeo-features/nuxeo-platform-collections/nuxeo-platform-collections-core/src/test/java/org/nuxeo/ecm/collections/core/test/CollectionPublishTest.java
@@ -47,7 +47,6 @@ import org.nuxeo.runtime.transaction.TransactionHelper;
 @RunWith(FeaturesRunner.class)
 @Features(CollectionFeature.class)
 @Deploy("org.nuxeo.ecm.user.center.dashboard")
-@Deploy("org.nuxeo.ecm.platform.tag")
 public class CollectionPublishTest {
 
     protected static final String TEST_FILE_NAME = "testFile";
