+++ nuxeo-features/nuxeo-platform-collections/nuxeo-platform-collections-core/src/test/java/org/nuxeo/ecm/collections/core/test/CollectionRightsTest.java
@@ -55,7 +55,6 @@ import org.nuxeo.runtime.test.runner.FeaturesRunner;
 @Deploy("org.nuxeo.ecm.platform.userworkspace.types")
 @Deploy("org.nuxeo.ecm.platform.query.api")
 @Deploy("org.nuxeo.ecm.platform.web.common")
-@Deploy("org.nuxeo.ecm.platform.tag")
 public class CollectionRightsTest {
 
     @Inject
