+++ nuxeo-features/nuxeo-platform-collections/nuxeo-platform-collections-core/src/test/java/org/nuxeo/ecm/collections/core/test/CollectionTestCase.java
@@ -50,7 +50,6 @@ import org.nuxeo.runtime.transaction.TransactionHelper;
 @Deploy("org.nuxeo.ecm.platform.userworkspace.types")
 @Deploy("org.nuxeo.ecm.platform.query.api")
 @Deploy("org.nuxeo.ecm.platform.web.common")
-@Deploy("org.nuxeo.ecm.platform.tag")
 public class CollectionTestCase {
 
     @Inject
