+++ nuxeo-features/nuxeo-platform-collections/nuxeo-platform-collections-core/src/test/java/org/nuxeo/ecm/collections/core/test/FavoritesAddRemoveTest.java
@@ -43,7 +43,6 @@ import org.nuxeo.runtime.test.runner.FeaturesRunner;
 @Deploy("org.nuxeo.ecm.platform.userworkspace.types")
 @Deploy("org.nuxeo.ecm.platform.query.api")
 @Deploy("org.nuxeo.ecm.platform.web.common")
-@Deploy("org.nuxeo.ecm.platform.tag")
 public class FavoritesAddRemoveTest {
 
     protected static final String TEST_FILE_NAME = "testFile";
