+++ nuxeo-features/nuxeo-platform-tag-service/nuxeo-platform-tag-core/pom.xml
@@ -65,6 +65,11 @@
       <type>test-jar</type>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.nuxeo.ecm.platform</groupId>
+      <artifactId>nuxeo-platform-collections-core</artifactId>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 
 </project>
